+++ rust/utils/run-locally/src/ethereum/multicall.rs
@@ -8,10 +8,3 @@ pub const DEPLOYER_ADDRESS: H160 = H160([
     0x05, 0xf3, 0x2B, 0x3c, 0xC3, 0x88, 0x84, 0x53, 0xff, 0x71, 0xB0, 0x11, 0x35, 0xB3, 0x4F, 0xF8,
     0xe4, 0x12, 0x63, 0xF2,
 ]);
-
-/// The Multicall3 contract address that is deployed in [`MULTICALL_SUPPORTED_CHAIN_IDS`]:
-/// [`0xcA11bde05977b3631167028862bE2a173976CA11`](https://etherscan.io/address/0xcA11bde05977b3631167028862bE2a173976CA11)
-pub const MULTICALL_ADDRESS: H160 = H160([
-    0xca, 0x11, 0xbd, 0xe0, 0x59, 0x77, 0xb3, 0x63, 0x11, 0x67, 0x02, 0x88, 0x62, 0xbe, 0x2a, 0x17,
-    0x39, 0x76, 0xca, 0x11,
-]);
